+++ sixel.c
@@ -1852,11 +1852,31 @@ decgra(struct terminal *term, uint8_t c)
         pan = pan > 0 ? pan : 1;
         pad = pad > 0 ? pad : 1;
 
+        if (likely(term->sixel.image.width == 0 &&
+                   term->sixel.image.height == 0))
+        {
+            term->sixel.pan = pan;
+            term->sixel.pad = pad;
+        } else {
+            /*
+             * Unsure what the VT340 does...
+             *
+             * We currently do *not* handle changing pan/pad in the
+             * middle of a sixel, since that means resizing/stretching
+             * the existing image.
+             *
+             * I'm *guessing* the VT340 simply changes the aspect
+             * ratio of all subsequent sixels. But, given the design
+             * of our implementation (the entire sixel is written to a
+             * single pixman image), we can't easily do that.
+             */
+            LOG_WARN("sixel: unsupported: pan/pad changed after printing sixels");
+            pan = term->sixel.pan;
+            pad = term->sixel.pad;
+        }
+
         pv *= pan;
         ph *= pad;
-
-        term->sixel.pan = pan;
-        term->sixel.pad = pad;
 
         LOG_DBG("pan=%u, pad=%u (aspect ratio = %d:%d), size=%ux%u",
                 pan, pad, pan, pad, ph, pv);